p4i integration Integrate 171137 from 4.0 to main: Fixes:    --warn Task:     none RevBy:    Trond AutoTest: none Details: MSVC and intel compilers on windows treats ints on the form 0xfe000000 as unsigned ints so combining these with ints causes warnings.
[git-p4: depot-paths = "//depot/qt/main/": change = 171138]

+++ src/gui/dialogs/qmessagebox.cpp
@@ -1005,12 +1005,12 @@ void QMessageBox::keyPressEvent(QKeyEvent *e)
     }
 #ifndef QT_NO_ACCEL
     if (!(e->modifiers() & Qt::AltModifier)) {
-        int key = e->key() & ~(Qt::MODIFIER_MASK|Qt::UNICODE_ACCEL);
+        int key = e->key() & ~((int)Qt::MODIFIER_MASK|(int)Qt::UNICODE_ACCEL);
         if (key) {
             QList<QPushButton *> list = qFindChildren<QPushButton *>(this);
             for (int i = 0; i < list.size(); ++i) {
                 QPushButton *pb = list.at(i);
-                int acc = pb->shortcut() & ~(Qt::MODIFIER_MASK|Qt::UNICODE_ACCEL);
+                int acc = pb->shortcut() & ~((int)Qt::MODIFIER_MASK|(int)Qt::UNICODE_ACCEL);
                 if (acc == key) {
                     emit pb->animateClick();
                     return;
